security/jwt.py
#DRF JWT
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from rest_framework_simplejwt.views import TokenObtainPairView

class MedicBackTokenObtainPairSerializer(TokenObtainPairSerializer):

    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        # Add custom claims
        token['username'] = user.username
        token['email'] = user.email

        return token


class MedicBackTokenObtainPairView(TokenObtainPairView):
    serializer_class = MedicBackTokenObtainPairSerializer
